+++ src/DataFrames.jl
@@ -144,9 +144,10 @@ include("abstractdataframe/io.jl")
 include("abstractdataframe/sort.jl")
 include("dataframe/sort.jl")
 
-include("deprecated.jl")
-
 include("other/tables.jl")
+include("other/names.jl")
+
+include("deprecated.jl")
 
 include("other/precompile.jl")
 precompile()