+++ src/groupeddataframe/groupeddataframe.jl
@@ -523,8 +523,6 @@ Base.getproperty(key::GroupKey, p::AbstractString) = getproperty(key, Symbol(p))
 
 Base.hash(key::GroupKey, h::UInt) = _nt_like_hash(key, h)
 
-_getnames(x::GroupKey) = parent(x).cols
-
 for eqfun in (:isequal, :(==)),
     (leftarg, rightarg) in ((:GroupKey, :GroupKey),
                             (:DataFrameRow, :GroupKey),
@@ -1019,8 +1017,5 @@ function _filter_helper_astable(gdf::GroupedDataFrame, nt::NamedTuple, f,
     return gdf[[f(mapper(i))::Bool for i in 1:length(gdf)]]
 end
 
-function Base.map(f, gdf::GroupedDataFrame)
-    Base.depwarn("Use of the map function on GroupedDataFrame is deprecated. " *
-                 "Use `[f(sdf) for sdf in gdf]` instead.", :map)
-    return collect(Base.Generator(f, gdf))
-end
+Base.map(f, gdf::GroupedDataFrame) =
+    throw(ArgumentError("using map over `GroupedDataFrame`s is reserved"))